Crate gpiosim

A library for creating controlling GPIO simulators on Linux platforms using the gpio-sim module.

Simulators can be built using the Builder.

For simple tests that only require one chip and control of line values, but not line info, the Simpleton provides a simplified interface.

The library controls the simulators via configfs and sysfs, so using it generally requires root permissions.

Structs

Bank

The configuration for a single simulated chip.

Builder

A builder of simulators.

Chip

A live simulated chip.

Hog

The configuration for a hogged line.

Sim

A live simulator of one or more chips.

Simpleton

A basic single bank/chip sim.

Enums

Direction

The direction, and for outputs the pulled value, of a hogged line.

Error

Errors returned by gpiosim functions.

Level

The physical value of a line.

Functions

builder

Start building a GPIO simulator.

simpleton

Build a basic single bank sim and take it live.

unique_name

Create a unique, but predictable, name for the simulator.

Type Definitions

Offset

The offset of a line on a chip.

Result

The result for gpiosim functions.